+++ ecs.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+use craft\ecs\SetList;
+use Symplify\EasyCodingStandard\Config\ECSConfig;
+
+return ECSConfig::configure()
+    ->withPaths([
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+        __DIR__ . '/ecs.php',
+        __DIR__ . '/rector.php',
+    ])
+    ->withSets([SetList::CRAFT_CMS_4]);